requirements.txt
python-telegram-bot==20.7
redis==5.0.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2 